Implement exercises
Exercise 1: declare 3 variables for name, city, and favNumber and
console.log() each like the following "My name is Omar"

Exercise 2: given the following string
"Java is different than JavaScript"
console.log(): the length of the string, the string in all uppercase,
extract the word javaScript using slice(), check if the string includes
the word javaScript

Exercise 3: given a variable myNumber that you pick the value of.
Check if the number is positive or negative. console.log a string like
the following "5 is a positive number"

Exercise 4: given the same variable myNumber check if the number is even
 or odd. HINT use the %

Exercise 5: given 3 numbers num1, num2, num3 console.log() the largest
number